requirements.txt
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
sqlalchemy==2.0.23
pymysql[rsa]==1.1.0
